requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0 